feat(gen-eval): modular checks (one module per check) + 4 new checks + per-check docs
Restructure: a single shared AST pass (src/parse.ts -> PackageFacts);
each check is now its own module under src/checks/ running as a pure
function over the facts — adding a check never adds a parse. The
accumulator verdict is computed in the facts because checks 3 and 5 key
their exemptions off it. src/checks/index.ts assembles GeneratorReport.

New checks:
- tostring-purity (pass/fail): no this.* assignments/mutations and no
  register-family calls inside toString bodies. Entire stock baseline
  passes — strong confirmation of the invariant.
- adhoc-tostring (pass/fail): no { toString: ... } object literals —
  these should be Snippets. Caught 3 stock offenders: gen-arktype
  (ArktypeObject, also an as-cast in the same expression),
  gen-reapit-form (2 sites), gen-reapit-graphql-client (1).
- as-casts (count, approval-required): as-expressions excluding
  , each site listed. Stock scatter: 0-3 per generator.
- registration-channels (informational): Driver-path insert* /
  defineAndRegister counts vs raw register({ definitions }) with
  hand-built Definitions. Baseline evidence: gen-kotlin-jackson-s's
  single raw site IS its whole output path; gen-shadcn-table's
  noExport columnHelper sibling is the candidate-legitimate case;
  gen-kotlin-sdk has 2 static-file sites.

Docs: docs/ now documents every check (one page per check + index):
what it asserts, why (tied to the generation model), how measured,
how to read results, known exceptions. Baseline regenerated with the
four new columns (pure/adhoc/as/raw-reg).
