+++ packages/gen-eval/src/types.ts
@@ -54,6 +54,85 @@ export type MethodDisciplineReport = {
   accumulatorExempt: { className: string; kind: ProducerKind; extraMethods: string[] }[]
 }
 
+export type CodeSite = {
+  file: string
+  site: string
+  line: number
+  text?: string
+}
+
+export type ToStringViolation = {
+  className: string | undefined
+  file: string
+  line: number
+  kind: 'assignment' | 'mutation' | 'register-call'
+  detail: string
+}
+
+export type RegistrationChannels = {
+  insertOperation: number
+  insertModel: number
+  insertNormalizedModel: number
+  defineAndRegister: number
+  rawDefinitionRegisters: CodeSite[]
+}
+
+// ---------------------------------------------------------------------
+// Facts — the output of the single shared AST pass (src/parse.ts).
+// Check modules are pure functions over PackageFacts.
+// ---------------------------------------------------------------------
+
+export type StringBucket = 'toString' | 'naming' | 'outside'
+
+export type ClassFacts = {
+  className: string
+  file: string
+  extendsName: string | undefined
+  extendsFactoryCall: boolean
+  methods: string[]
+  lines: number
+  containerProps: string[]
+  mutatorMethods: string[]
+}
+
+export type FileFacts = {
+  file: string
+  classes: ClassFacts[]
+  projectionBaseConsts: string[]
+  snippetImports: string[]
+  peerProjectionImports: string[]
+  helperFunctions: string[]
+  usesDefineAndRegister: boolean
+  usesFindDefinition: boolean
+  stringSites: Map<string, { count: number; chars: number; bucket: StringBucket }>
+  toStringViolations: ToStringViolation[]
+  adHocToStringSites: CodeSite[]
+  asCastSites: CodeSite[]
+  insertCalls: {
+    insertOperation: number
+    insertModel: number
+    insertNormalizedModel: number
+    defineAndRegister: number
+  }
+  rawDefinitionRegisters: CodeSite[]
+}
+
+export type PackageFacts = {
+  dir: string
+  packageName: string | undefined
+  denoJsonParseError: boolean
+  fileCount: number
+  files: FileFacts[]
+  classes: ClassReport[]
+  producers: ClassReport[]
+  helperFunctions: string[]
+  accumulator: AccumulatorReport
+}
+
+// ---------------------------------------------------------------------
+// The assembled per-generator report (one entry per check module).
+// ---------------------------------------------------------------------
+
 export type GeneratorReport = {
   generator: string
   dir: string
@@ -68,4 +147,8 @@ export type GeneratorReport = {
   topLevelProjection: { pass: boolean; exempt: boolean }
   accumulator: AccumulatorReport
   producerSizes: { bucket: number; count: number }[]
+  toStringPurity: { pass: boolean; violations: ToStringViolation[] }
+  adHocToString: { pass: boolean; sites: CodeSite[] }
+  asCasts: { count: number; sites: CodeSite[] }
+  registrationChannels: RegistrationChannels
 }
